+++ packages/web/src/app/doc-pages/dev/v2-playground/v2-playground.component.ts
@@ -81,4 +81,7 @@ export class v2PlaygroundComponent {
     const values = [...this.deletableChipsList];
     this.deletableChipsList = values.filter((value) => value.value !== event);
   };
+  datepickerOnError = (error: string): void => {
+    console.log('Datepicker error: ', error);
+  };
 }